Update readme.md
# Encrypter/Decrypter

This is a project developed for Challenge 1 of Alura+Oracle. The objective of the project is to create an encrypter and decrypter for messages using HTML, CSS, and JS.

## Functionality

The encrypter/decrypter uses a simple algorithm that replaces each vowel in the original message with a combination of letters defined in a key object. For example, the letter "a" is replaced with "ai", the letter "e" with "enter", and so on.

In addition, the project includes a language selector that allows users to change the language of the texts in the interface to one of 11 possible languages: Spanish, Portuguese, English, French, German, Italian, Japanese, Korean, Russian, Arabic, and Chinese.

To encrypt a message, the user must enter the original message and select the language in which they want the encrypted message to appear. The algorithm replaces each vowel in the original message with the corresponding combination of letters defined in the key object and displays the encrypted message in the selected language.

To decrypt a message, the user must enter the encrypted message and select the language in which it appears. The algorithm performs the inverse process, replacing each combination of letters in the encrypted message with the corresponding vowel and displaying the original message in the selected language.

## How to use

To use the encrypter/decrypter, simply open the `index.html` file in your web browser. Enter the message you want to encrypt or decrypt in the corresponding text box and select the language in which you want the resulting message to appear. Press the "Encrypt" or "Decrypt" button as appropriate. The resulting message will appear in the text box below.

## Credits

This project was developed as part of Challenge 1 of Alura+Oracle, an educational bootcamp organized by Oracle Next Education. The technologies used in the project are HTML, CSS, and JS.

## License

This project is distributed under the MIT License. Please see the `LICENSE` file for more information.